feat(streamcreate): add button styled <Link> element for a cancel button
NOTE: Implemented as a styled react-router-dom <Link> and not a <button>. Observe two methods of navigatio n. With Link there is no behaviur code. (It is implicit). With a <button> we need a click handler code, and then a hisory obj to navigate. So, Link seems to make sense for pure navigation, button for complex behaviour